local_vendors_marketplace/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
